feat: setdata and update idb
