+++ providers/ic/src/api/data.api.ts
@@ -120,53 +120,40 @@ export const getData = <T, D>({
   });
 };
 
-export const setData = <T, D>({
+export const setData = async <T, D>({
   key,
   data,
   id,
-  actor = undefined,
-  log
+  actor = undefined
 }: {
   key: string;
   data: D;
   id: string;
   actor?: DataBucketActor;
-  log?: LogWindow;
 }): Promise<T> => {
-  return new Promise<T>(async (resolve, reject) => {
-    try {
-      log?.({msg: `[set][start] ${key}`});
-      const t0 = performance.now();
+  const dataActor: DataBucketActor = actor || (await getDataActor());
 
-      const dataActor: DataBucketActor = actor || (await getDataActor());
+  // Used to prevent edge case
+  const now: Date = new Date();
 
-      const now: Date = new Date();
-
-      // We update the data with the current updated_at time
-      // The canister will check if the updated_at date is equals to the entity timestamp otherwise will reject the update to prevent overwrite of data if user uses multiple devices
-      await dataActor.set(key, {
-        id,
-        data: await toArray<D>(data),
-        created_at: toTimestamp((data as unknown as {created_at: Date}).created_at ?? now),
-        updated_at: toTimestamp((data as unknown as {updated_at: Date}).updated_at ?? now)
-      });
-
-      const t1 = performance.now();
-      log?.({msg: `[set][done] ${key}`, duration: t1 - t0});
-
-      const result: T = {
-        id,
-        data: {
-          ...data,
-          updated_at: now
-        }
-      } as unknown as T;
-
-      resolve(result);
-    } catch (err) {
-      reject(err);
-    }
+  // We update the data with the current updated_at time
+  // The canister will check if the updated_at date is equals to the entity timestamp otherwise will reject the update to prevent overwrite of data if user uses multiple devices
+  await dataActor.set(key, {
+    id,
+    data: await toArray<D>(data),
+    created_at: toTimestamp((data as unknown as {created_at: Date}).created_at ?? now),
+    updated_at: toTimestamp((data as unknown as {updated_at: Date}).updated_at ?? now)
   });
+
+  // TODO: actor returns value to update timestamp
+
+  return {
+    id,
+    data: {
+      ...data,
+      updated_at: now
+    }
+  } as unknown as T;
 };
 
 const getDataActor = async (): Promise<DataBucketActor> => {
