+++ providers/ic/src/providers/feed/feed.providers.ts
@@ -9,9 +9,9 @@ import {
   Meta
 } from '@deckdeckgo/editor';
 import {Identity} from '@dfinity/agent';
-import {setData} from '../../api/data.api';
 import {_SERVICE as FeedActor} from '../../canisters/feed/feed.did';
 import {_SERVICE as StorageBucketActor} from '../../canisters/storage/storage.did';
+import {setData} from '../../services/data.services';
 import {EnvStore} from '../../stores/env.store';
 import {toNullable} from '../../utils/did.utils';
 import {createFeedActor} from '../../utils/feed.utils';
