+++ providers/ic/src/publish/doc.publish.ts
@@ -1,5 +1,5 @@
 import {Doc, DocData, docPublishData, DocPublishData, PublishData} from '@deckdeckgo/editor';
-import {setData} from '../api/data.api';
+import {setData} from '../services/data.services';
 import {EnvStore} from '../stores/env.store';
 import {
   initIndexHTML,
